allow chat panel to load even if unauthenticated
Prior to my changes, `ConfigFeaturesSingleton` would _ignore all GraphQL
errors_ and report that chat is enabled in that event. As it turns out,
we were relying on that behavior for the login panel to appear: because
the login panel /is/ the chat panel, chat 'must be enabled' in order for
the login panel to be displayed.

In this commit, I elevate that logic up to where we actually need to do
that. Chat disablement as a feature needs some work, see https://linear.app/sourcegraph/issue/PRIME-409

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>