Autocomplete: decrease requests timeout (#4813)
Reduce the timeout for autocomplete requests from 15 to 7 seconds. This is the "reasonable" value for this timeout. Based on data for the last 30 days, 0.07% of all autocomplete requests took longer than 5s. I can bump it to 7s to make this even lower — 0.03%. We expect this to happen during incidents only or on a very bad connection where autocomplete is not usable.