Several fixes for Hotkey Inlay Hints (#1895)
Based on user feedback from GTM and Product, discussions with
@danielmarquespt , and my own testing, I've made the following round of
fixes to the hotkey hint feature, which shows an inlay with "Ctrl + Alt
+ ↵ to Edit" at the end of your selections in the editor window:

1. There is a new Preferences option for disabling the feature:
<img width="653" alt="image"
src="https://github.com/sourcegraph/jetbrains/assets/613744/8a78f2b1-c6d5-4a9c-902b-687d7d2e236d">

2. The hint no longer draws if the selection is only on one line.
- this takes care of many annoying cases, such as Find/Replace lighting
it up everywhere
 
3. The hint now draws below the last line of the selection.

This makes it less intrusive, IMO, as it's out of the way of the cursor.

I think it's also more intuitive. Several users claim that it "used to
work that way". It never actually worked that way; it always drew at the
end of the last line of the selection. But now it does work that way.

The one downside is that the hint can sometimes draw rather far from the
selection, appearing to be two lines below. We will see how people
respond to it.

Coincidentally and usefully, this change also fixes CODY-2541. Getting
the hint inlay out of the way of the cursor removed the interference
with fine-tuning the selection.

These changes also appear have to fixed CODY-2748, because I reported it
and I can no longer reproduce it. It used to be 100% reproducible, so
I'm going to mark it closed until it resurfaces.

fixes CODY-2541
fixes CODY-2748

## Test plan

This requires UI testing to verify with actual integration tests.
I reviewed the test plan, and these changes do not affect the plan.