check:manifest uses committed schema cache
This step was taking 7s for me, it now takes 1s. I decided to commit the
files since this will save time on CI as well, although likely not as
significant due to the servers likely being closer to schemastore
servers.

I removed the recently added OFFLINE support, since this step now runs
offline.

Note: In a previous PR I was worried about the complexity of doing this.
However, Cody edit managed to just generate really start for this and I
believe the final commit is simple enough and is worth the speedup.

Test Plan: "pnpm -C vscode run check:manifest"