report to user when Experimental Generate Unit test fails
If a prompt is not generated now, we will always throw an error.
Previously those errors would only go to Cody Output and be hard to find
for the user. So we wrap the call with a try catch so we can report the
errors.

Note: Ideally we should only show the button if we have an active
document (the easiest failure case). That is for the future.

Test Plan: manually tested with no active document and with an active
document.