edit: quick pick option for full file for range
I find myself quite often wanting to run edit on the whole file.
Currently I select the whole file and then use that selection range.
However, it seems useful to include it as one of the quick pick options.

Note: I included this option in any place we could include a selection.
I did not exclude this option if the full file was already selected,
since that seemed like extra complications.

This idea is kind of like Steve's CHOP idea, except you skip needing to
use copy + paste from chat as well as getting built in diff view.

Test Plan: triggered an edit and selected the full file as the range.