Agent: add support for Visual Studio IDE in AgentWorkspaceConfiguration
This change adds support for the Visual Studio IDE in the AgentWorkspaceConfiguration class. A new case has been added to the `getIDE()` method to handle the 'visualstudio' configuration value and return the corresponding CodyIDE.VisualStudio enum value.

Additionally, the CodyIDE enum in the shared configuration.ts file has been updated to include the VisualStudio value.