Chat: use Context Preamble by default (#4912)
Follow up on
https://linear.app/sourcegraph/issue/CODY-2507/evaluate-claude-35-sonnet-as-new-default

As mentioned in the linked issue, we were able to verify the context
preamble has improved chat quality. This PR is to remove the
CodyChatContextPreamble feature flag so that the Context Preamble is
always used by default. This also applies to Enterprise users as they do
not have access to feature flags.

The CodyChatContextPreamble feature flag has been removed, and the
context preamble is now always enabled by default in the PromptMixin.
This simplifies the code and removes the need for the feature flag.

BREAKING CHANGE: The CodyChatContextPreamble feature flag has been
removed. The context preamble is now always enabled by default when
context is used.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Green CI, as all existing features should remain unchanged. All our
tests should not affected by this change because the same preamble was
enabled for all tests.