chat: template chips for prompts (#4864)
We introduce experimental rendering of template chips in prompts. This
is only turned on for "Generate Unit Tests (experimental)" button, but
in the future can be for everyone.

Right now the template syntax is "{{ placeholder text }}". I am assuming
this will change before it is available to other users. For example
right now custom commands have hardcoded "{repo}"/etc templates. When we
ship prompt templates, we likely want something nice here that mixes the
ideas of values that cody can automatically infer vs what users need to
input.

The UX for this is quite barebones right now, mainly so that we can
follow-up and make it better. This PR is mostly about the plumbing and
making sure everything fits together.

One thing to note is the serialization of editorState now changes due to
the new template input chips. We have made this backwards compatible.
Additionally, because this feature is experimental, we have taken care
to only serialize in the new format if the editor state has template
chips in it.

### Test Plan

We added a fixture for template chips. We then did manual testing inside
of storybook via the fixture's transcript template. Additionally manual
testing of the generate unit test feature.


![image](https://github.com/user-attachments/assets/b1c9b42f-5f16-47d3-bc1e-78d48c02d946)

Fixes CODY-2800