Exclude context chips from the query text (#4920)
There is a bug in our chat context fetching that causes embeddings to
not work at all and makes symf perform worse.
We pass the full PromptString to these retrievers, which means that we
include all context mentions. For example, opening a chat in the cody
repo, @mentioning src/context.ts and writing "what is symf" will make
the query look like `cody src/context.ts what is symf?`. The reason is
that we include everything in the chat text box as the query, including
the context chips.

Credits to @olafurpg for the fix - we now exclude the context chips from
the query text which fixes the issue. Should we cut a patch release to
get this out ASAP?

## Test plan
Tested locally - bfg and symf now return correct results.