feat(cached-retriever): handle non-file scheme URIs in openTextDocument (#4942)
CLOSE
https://linear.app/sourcegraph/issue/CODY-2400/bug-cody-creates-unsved-files-keep-showing-back-after-closing

- Adds a check to `openTextDocument` to return `undefined` if the URI is
not a file scheme, which includes untitled files
- Adds a similar check in `onDidChangeTextEditorSelection` to skip
handling selection changes for non-file scheme URIs


This ensures the cached retriever does not try to open or handle text
documents that are not file-based, including untitled files. As a
follow-up task for the team, we should look into improving the cache
retriever logic by tracking closed and removed files, but this fix the
issues we are seeing with untitled files.

The video below shows that the closed untitled files keep getting
re-opened. The added ogging step (during debug) before
`openTextDocument` is run shows `openTextDocument` is the issue as
opening a non-existing file will create them:


https://github.com/user-attachments/assets/1b170d0e-b213-4740-99d5-1c19dcbe6a50

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Try reproducing the issue detailed in
https://github.com/sourcegraph/cody/issues/4581 to verify autocomplete
no longer re-open closed & unsaved files.

```
Disable all installed extensions (does not seem to be strictly necessary for me, but I include this to show it's not an interaction with another third party extension)
Create a new workspace with a main.py in it
Press Ctrl+N to create a new untitled tab
Close the untitled tab, and notice that it stays closed as expected
Enable Cody
Press Ctrl+N to create a new untitled tab
Close the untitled tab, and notice that it stays closed
Type something in the main.py file to trigger an inline autocomplete suggestion (no need to accept it)
Press Ctrl+N to create a new untitled tab
Close the untitled tab
```