graphql: use ETIMEDOUT for request timeouts (#4960)
The EHOSTUNREACH error was misleading, since we only know the request
timed out and not that the host was unreachable. I believe we used
EHOSTUNREACH since the original thing we where debugging was unexpected
behaviour when using a proxy. This was introduced in
https://github.com/sourcegraph/cody/pull/4243

ETIMEDOUT isn't exactly correct since we aren't getting a TCP layer
timeout. However, I am using the errno approach so the isNetworkError
code still works.

Note: the autocomplete code uses specific error types to encode these
behaviours which we should switch to. However, that is a bigger change
and I wasn't sure how it would work given the RPC stuff between
extension worker, webview, etc.

Test Plan: CI

Part of CODY-2959