guardrails: set timeout for graphql request (#4964)
In https://github.com/sourcegraph/cody/pull/4243 a default timeout was
added to all GraphQL requests of 6 seconds, which is lower than the
timeout for Guardrails. This lead to the timeout we had hardcoded to
never triggering.

This adjusts the implementation to instead specify the timeout for the
API client and rely on its error message. An alternative implementation
here we could inspect the error thrown translate it into the Timed out
message. However, I would rather do that as a follow-up since this won't
change the current presentation, but rather just give us more time.

Test Plan: Just relying on CI and typescript to tell me if I messed up.

Fixes CODY-2959