Autocomplete: set the default anthropic model only for certain cases (#4982)
- Fixes the issue when enterprise instances use the non-deployed
Anthropic model as a default value. Now we set the default Anthropic
model only for DotCom users and only in cases where the autocomplete
provider is set to Anthropic in local settings. It keeps [the original
issue](https://github.com/sourcegraph/cody/issues/4639) fixed and also
addresses the enterprise problem.
- Slack investigation
[here](https://sourcegraph.slack.com/archives/C07CVL78MDF/p1721694416152009).