Chat UI: add webview type to webview configuration (#4951)
This allows the chat UI to render based on the webview type. Example:
don't show nav bar in editor chat view.

- Adds a new `webviewType` property to the
`ConfigurationSubsetForWebview` interface in `protocol.ts`
- Updates the `ChatController` to set the `webviewType` based on the
webview panel or view type
- Conditionally renders the `TabsBar` component in `App.tsx` based on
the `webviewType`. It will not show on editor view.

This change allows the webview to adjusts the UI displayed in either the
sidebar or the editor accordingly.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->


![image](https://github.com/user-attachments/assets/cb9ace5c-96bd-4b58-9aaa-c9fcf7a9349f)