Autocomplete: Check completions are _still_ visible before logging them as suggested and `read` (#4984)
## Description

This PR fixes an issue where we would incorrectly log a suggestion as
"suggested" even though it wasn't actually suggested by our intended
criteria.

### Problem

When we determine a completion is visible, we currently "chuck it over
the fence" to the completion logger, and assume it is still visible
after 750ms. There are some cases where that might not happen, for
example if the user triggers a new completion, or deletes some
characters then we explicitly will not log it. However we did not handle
if the user moves their cursor away from the completion, or leaves the
document _before_ those 750ms have passed. In those cases the completion
will certainly not be visible in VS Code, but will be marked as visible
in our analytics.

Example of how this can happen:
1. Cody generates a completion
2. We immediately start a function to wait 750ms before logging the
completion as suggested (so the user has chance to read it, otherwise
our metrics may be affected by completions that the user didn't even
acknowledge)
3. User immediately moves cursor away from completion, forcing VS Code
to hide it
4. After 750ms we log the completion, regardless of the fact that VS
Code hid it.

### Solution

We already check `isCompletionVisible` before determining if we should
suggest it, we need to do the same after 750ms but using the latest
`activeTextEditor` and `activeTextEditor.selection.active` cursor
position.

## Test plan

Tested locally by replicating the example above.

Unit tests