chat: factor out helpers over SerializedLexicalNode  (#4996)
We implemented a pre-order walk using a queue 4 times. This factors out
the common logic into a function `forEachPreOrder`. Additionally in two
of the cases we did filtering, so we also introduce the helper
`filterLexicalNodes`.

Note that we use recursion instead of a queue. We don't have deep trees
to recurse. I believe previously we used a queue just because it was
more convenient due to the function signature. But now it is easier (and
clearer I think) to use recursion.

Test Plan: call sites are covered by tests.