helper for webview to call extension (ChatController) API and get a result (#5006)
This is a useful pattern that we use for fetching @-mention context.
This helper makes it easy to add more. In the future we can adopt a real
RPC protocol for exthost<->webview communication, and this helper will
make that eventuality easier.

## Test plan

e2e tests. Also try a lot of @-mentions in chat and see if any behavior
changes.