guardrails: experimental configuration of timeouts (#5020)
We are currently experiencing long response times from guardrails
servers. We have a hardcoded timeout of 10s. Until the issue is resolved
we want to provide a way to workaround the timeout. It is marked
experimental since I don't expect this setting to remain in the config
forever. Instead when we invest more in the guardrails product I would
expect this to live server side (along with other parts of the
configuration for it).

Note the configuration key is
"cody.experimental.guardrailsTimeoutSeconds"

Test Plan: hardcoded guardrails on to test against s2. Then configured
the following and tested guardrails by asking chat to generate a sort
implementation of more than 10 lines:
- unconfigured :: default timeout of 10s used
- 0.1 timeout :: timed out
- 40 :: no timeout