@-mention git unmerged changes vs. main or uncommitted changes (#5034)
This is an experimental new context provider that lets you @-mention Git
information:

- Unmerged changes (`git diff origin/main...HEAD` and `git log
origin/main..HEAD`)
- Uncommitted changes (`git diff HEAD`)

It can be used to get "pre-reviews" of your diff.

To use it, enable the `git-mention-provider` feature flag (which is
enabled on s2 and for all `sourcegraph` org members on dotcom).

<img width="508" alt="image"
src="https://github.com/user-attachments/assets/fbc31259-f329-484a-ba94-f78f33a26a96">


## Test plan

Run in a debug VS Code. Try @-mentioning it and ensure it gets the
context in the response.