noodle: remove generate unit test button
Going forward this will just be a prompt we can test as part of the
prompt library. This attempts to rip out all the code related to the
feature, except the stuff around template chips/etc which I expect to be
used by prompt library.

Test Plan: pnpm build and pnpm test