don't show file 2x in mention menu
In the chat @-mention menu, often the current file was shown twice: once at the top because it's part of initial context, and once below because it's an open tab. This caused a glitch where it would be highlighted in 2 places in the mention menu and would prevent navigating down with the arrow key. Now it's only shown once.

Fix https://linear.app/sourcegraph/issue/CODY-2749/mention-menu-glitches-when-you-try-to-navigate-with-the-updown-arrows