don't show file 2x in mention menu (#5057)
In the chat @-mention menu, often the current file was shown twice: once
at the top because it's part of initial context, and once below because
it's an open tab. This caused a glitch where it would be highlighted in
2 places in the mention menu and would prevent navigating down with the
arrow key. Now it's only shown once.


![image](https://github.com/user-attachments/assets/8a276c5c-5228-441f-8f74-f0c23d16f3e8)


Fix
https://linear.app/sourcegraph/issue/CODY-2749/mention-menu-glitches-when-you-try-to-navigate-with-the-updown-arrows

## Test plan

Open mention menu and confirm only shown once.

## Changelog

- Fixes a bug where the same file was sometimes shown twice in the
@-mention menu, causing a glitch in arrow-key navigation.