Chat: boost `@-file` mentions for open documents
Previously, when using `@-file` mentions for common filenames in the
repo, the ordering of the results wasn't personalised meaning the top
result could be a random file (for example, a short file path) instead
of a file that the user has open in their editor. This PR adds basic
customization where we boost candidates that the user has open in their
editor.