Chat: boost `@-file` mentions for open documents (#5068)
Previously, when using `@-file` mentions for common filenames in the
repo, the ordering of the results wasn't personalised meaning the top
result could be a random file (for example, a short file path) instead
of a file that the user has open in their editor. This PR adds basic
customization where we boost candidates that the user has open in their
editor.


## Test plan

Open local build of cody in the sourcegraph repo. Open a file named
`openai.go` which is not in cody-gateway, open another file named
something else. Open chat and type `@openai.go` and confirm that the top
result is the `openai.go` that you have open instead of the one in
cody-gateway. Confirm that you can optionally type a space character and
`cody-gateway` to narrow the filter to boost the cody-gateway result to
the top.

Before
![CleanShot 2024-07-31 at 11 36
09@2x](https://github.com/user-attachments/assets/f1767c73-802f-4d61-9223-8ef952e398dc)

After 
![CleanShot 2024-07-31 at 11 36
45@2x](https://github.com/user-attachments/assets/302af5f2-9e89-4413-abd7-be9333baeb10)

Workaround
![CleanShot 2024-07-31 at 11 37
30@2x](https://github.com/user-attachments/assets/5ec756ef-540f-4103-8d4c-f3d9803ea524)

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog
When using `@-file` mention in Chat, open documents are now ranked at
the top of the fuzzy matcher list.

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->