Code Actions Support (Ask Cody to Fix) (#1921)
Add the ability for the Agent to provide CodeActions. In practice this
means "Ask Cody to Fix" is now supported as a QuickFix on errors shown
in JetBrains.

https://github.com/user-attachments/assets/f3bad3fd-cd67-427c-baa6-449a4c0dfd48

## FAQ

**Q: What errors are supported?**
A: Any errors shown in the JetBrains UI. Cody doesn't do any code
analysis, we simply pass the issues that JetBrains is showing along.
These Issues (or Inspections as they're known) can come both from
JetBrain's language support, external Plugins as well as custom
Inspections.

**Q: Can Cody explain the error like in VSCode?**
A: Not yet, but the foundational support is there now so it will be easy
enough to add soon.

Fixes CODY-2788, CODY-2830

## Test plan
- Verified that existing unit & integration tests work with Protocol
migrated protocol classes
- Manually verified that "Ask Cody to Fix" command is correctly shown
and invoked. Most of the logic remains on the Agent.
- Isolated changes on FixupSession to FixupSessionV2 to not introduce
bugs in FixupSessionV1 dependent codepaths which currently have low test
coverage.