openctx: call disposable when replacing openctx controller (#5071)
We call the setOpenCtx everytime the config or authed account changes.
Previously we correctly would replace the OpenCtx client when that
happened, but we did not call dispose on the old client or the
observable disposables created by vscode-lib.

Test Plan: at mentioned remote repos and got s2 repos. Switched to
dotcom and got dotcom repos instead. So switching still works.