Experimental: fetch context from both remote and locally modified source files (#5067)
This modifies behavior when the `cody.internal.serverSideContext`
setting is true. If this setting is not set, there should be no behavior
changes.

- Remote context is fetched from Sourcegraph, the latest HEAD revision
of the repository (which is how enterprise context currently works)
- Local context is fetched from files changed since the merge base of
the current local working state and the upstream of the current branch.
If the upstream doesn't exist, then the changes since last local commit
are used.
- Context from the locally edited files is removed from the remote
context to avoid duplication/staleness.
- Currently the local context is limited to 5 entries and concatenated
with the remote context. In the future, we should rerank these context
items together.