openctx: merge in "openctx.providers" from Sourcegraph instance
If "openctx.providers" is configured in a users settings on a
Sourcegraph instance, it will merge that in. This allows admins and
users to set the configuration and automatically have it included in
Cody.

The implementation prefers the vscode settings for a provider, but will
merge the list of remote and local providers. This is hidden behind the
noddle configuration so we can dogfood this.

Test Plan: added a provider to my sourcegraph settings and it appeared
in Cody.