use streaming RPC API exthost<->webview using AsyncGenerator (#5083)
Now, the webview can call into API methods exposed by the ChatController
(extension host), and those API methods can return a stream of values.
This lets the webview "observe" things in the extension host.

Fixes several issues:

- OpenCtx @-mention providers now load immediately when each is ready,
without waiting for all to be ready. We had a 5-second hardcoded delay
for providers, which meant that in many cases the first chat session
would not have any OpenCtx providers. Also, if you change your
configuration, the new providers appear immediately in all chats.
- Feature flags used in the webview with `useFeatureFlag` now load
properly. This fixes an issue where the Prompts dropdown was not
displaying, even for users for whom it was enabled.
- Context items now stream in, which means that an OpenCtx provider can
emit multiple values as it loads more and more data.
- Webview<->exthost API calls are now aborted automatically (when using
the `useAsyncGenerator` React hook).

## Test plan

CI. Also a lot of manual usage of @-mentions and the Prompts dropdown,
since those are the parts that use this new API.