Fix integration tests (#1975)
### Changes

This PR fixes few problems:

* We were restarting agent without waiting for it to finish
* We were not properly refreshing external file changes to VFS which
sometimes was leading to having old files in memory

It's also worth noting that `BasePlatformTestCase` seems to be in face
reusing the same project for all tests, even if it does some minor
cleanup on it. I think this is not ideal for our case, and instead we
should try to either:

* reuse the same project and do not restart agent every time
* have full project isolation

Such in-the-middle solution as we have now does not seem to be
beneficial to me. But we can try to change that later, especially in
case this PR would not some all integration tests instabilities.

## Test plan

N/A

---------

Co-authored-by: Mikołaj Kondratek <mik.kondratek@gmail.com>