pass legacy context fetchers directly into ContextFetcher
Note that there is a divergence in behavior, because these legacy context fetchers are used
in the indexed context code path, whereas before they were used for fetching overall
context. So the code is in an intermediate state. The next steps is to rewrite
ContextFetcher.fetchContext to take HumanInput as an argument, and then to ensure the right
things are set for both the PLG and enterprise case.