Fall back to symf when remote indexed context is not available (#5108)
Implements part of
https://linear.app/sourcegraph/issue/AI-160/unify-context-fetching-code-paths

This change rolls in nearly all existing local PLG context into the
`ContextRetriever` class, which will be the new focal point for
retrieving unified context from both client and server. The new context
retrieval is enabled with the `cody.internal.serverSideContext` flag.

The context incorporated in this change includes:
* Explicit symbol and file mentions
* OpenCtx context
* "Priority" context, such as the README when the "include README"
heuristic is triggered
* Indexed symf results (this is used as a fallback when server-side
context does not exist for a given repository)

The context that has *not* been incorporated by this change:
* Local embeddings

The new context construction code path and the legacy context have been
factored cleanly into `legacyComputeContext` and `_computeContext`
methods within `ChatController` to make it clear what the delineation
is.