show tab labels on wide screens (#5124)
If the screen is 768px wide or wider, then the Cody chat panel tabs will
show text labels. This means that the standalone Cody Web chat screen
will show text labels for most users, which fixes
https://linear.app/sourcegraph/issue/CODY-3171/history-tab-is-hard-to-find-on-cody-web.



https://github.com/user-attachments/assets/7790d9ff-af8e-4e6f-929a-ea3210be1f2d



## Test plan

Make screen (or containing panel) narrow and wide. When it's wide,
confirm tooltips are NOT shown but tab titles are visible. When narrow,
confirm tab titles are NOT shown but tooltips are.