CodeActions: Explicitly disable ExplainAction (#1994)
We were not checking action types correctly. This was caused by
confusion during the PR review where the flag statement was supposed to
both check that the action was known (so we can disable unknown
actions), but then separately flag the ExplainAction (which is known) to
be disabled as well.

Now by separating the flags the intent is hopefully clearer and we
ensure that ExplainActions are explicitly disabled while they're pending
implementation.

Fixes [CODY-3168](https://linear.app/sourcegraph/issue/CODY-3168) /
[#1989](https://github.com/sourcegraph/jetbrains/issues/1989)

## Test plan
Manually verified that Explain no longer shows up.