misc. Cody Web usability improvements (#5129)
This makes it so that when you go to `/cody/chat` in Cody Web, it
autofocuses the message input for a new chat. This is the common desired
user behavior.


- Always start a new chat in Cody Web. Now that the history tab is
easily visible, it's easy for users to open other chats in the Cody chat
sidebar alongside a file in Sourcegraph.
- Even when you're viewing a file and open the Cody chat sidebar, it
opens a new chat. The idea of "keep a running chat per-file" was a cool
idea, but not necessary since we now have the history tab.
- Open the Prompts & Commands and Chat Help info boxes initially (in
both Cody Web and in editors). If the user closes them, that state will
persist. We had made them hidden by default to make them less noisy, but
now I think persisting the closed state is sufficient to make it less
noisy.


## Test plan

In Cody Web (sourcegraph/sourcegraph pnpm link'd), test that each time
you load Cody standalone chat (`/cody/chat`), it creates a new chat.