Agent: Revert breaking protocol change (#5128)
There was a unintentional protocol change to a legacy `getUserContext`
message where it didn't trigger the associated `userContextFiles`
message. This message was still being used by Jetbrains causing
[CODY-3196](https://linear.app/sourcegraph/issue/CODY-3169/sourcegraphjetbrains1990-jetbrains-file-tagging-is-not-working)

This reverts the change and ensures that the correct messages are
triggered. As a follow up
[CODY-3172](https://linear.app/sourcegraph/issue/CODY-3172/cascading-messages-arent-type-checked)
will ensure that these class of issues are properly (type-)checked and
tested.

## Test plan
- Current tests pass
- Verified locally that with the change JetBrains once again shows
context files
- Deferred testing of the actual testing of the message handler until
CODY-3172