Remove query rewrite experiment (#5139)
This removes the Query rewrite experiment (under noodle flag) I launched
during the Noodle project to check if LLM-based query rewrite improves
the context fetching. Due to limitations on the evaluation dataset side,
I wasn't able to test it.

Cleaning it up for now. 

## Test plan

Test Cody Chat and make sure the context fetching is not affected. 
## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->